Update README.md
Are you looking to step into the world of crypto trading or diversify your trading platforms? If yes, then WEEX might be the opportunity you've been waiting for. This May, WEEX is offering exclusive welcome bonuses for new users who register using the **referral code (ckk0)**. Whether you're a beginner or an experienced trader, this promotion is a great chance to explore the platform while enjoying free rewards.